refactor: #2 비즈니스 예외 처리를 GeneralException 기반 구조로 개선
