+++ src/main/java/or/sopt/houme/global/api/BusinessException.java
@@ -1,0 +1,10 @@
+package or.sopt.houme.global.api;
+
+import lombok.Getter;
+
+@Getter
+public class BusinessException extends GeneralException {
+    public BusinessException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
